Support displaying the context trace as well
This introduces new settings and code to configure the display such that
we can display either the scope, the scope context stack, or both,
depending on the setting.

The display of the context stack requires build 4087 or better since
that is the build where the API required to do this was introduced.